// web/src/api/schemas/recipesRecipeResponse.ts
/**
 * Generated by Orval
 * Do not edit manually.
 * Domicile API
 * Domicile's REST API
 */
import type { GithubComDomicileappDomicileInternalDbListRecipeIngredientsRow } from './githubComDomicileappDomicileInternalDbListRecipeIngredientsRow.ts';
import type { GithubComDomicileappDomicileInternalDbListRecipeInstructionsRow } from './githubComDomicileappDomicileInternalDbListRecipeInstructionsRow.ts';

export interface RecipesRecipeResponse {
  cook_time?: string;
  created_at?: string;
  deleted_at?: string;
  id?: number;
  ingredients?: GithubComDomicileappDomicileInternalDbListRecipeIngredientsRow[];
  instructions?: GithubComDomicileappDomicileInternalDbListRecipeInstructionsRow[];
  name?: string;
  notes?: string;
  nutrition?: string;
  photo_url?: string;
  prep_time?: string;
  servings?: string;
  short_description?: string;
  source?: string;
  updated_at?: string;
}